Add inputs argument handling and actual exec of compiler.
The inputs are added to the command in the order they appear. Like the
regular options. This also now executes the composed command. So far
this uses the std::system call which is not precise as it goes through
platform shell interpreter. At some point I'll debug why the fork+exec
segfaults.