Add first vendor specific option.
Add the GCC vendor specific cpp_dialect as first test case for vendor
options. To make the "++v:<vendor>:<option>" work with Clara did some
patches to add customization of the option parsing. And now we have both
the "++" option prefix and correct vendor syntax. We now have Clara as a
submodule of the fork in bfgroup.